data-getter/requirements.txt
scrapy>=2.11.0
requests>=2.31.0
lxml>=4.9.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0